requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-text-splitters>=0.0.1
langchain-core>=0.1.0
langchain-classic>=1.0.0
pypdf2>=3.0.1
pdfplumber>=0.10.3
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
openai>=1.12.0
python-dotenv>=1.0.0
chromadb>=0.4.22
tiktoken>=0.5.2
numpy>=1.24.0
pandas>=2.0.0
altair>=5.0.0
Pillow>=10.0.0
